ai_backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.43.0",
]
